feat: implement TrieWriter and StorageTrieWriter
